+++ policy_sentry/command/query.py
@@ -76,40 +76,54 @@ def action_table(name, service, access_level, condition, wildcard_only, fmt):
             for serv in all_services:
                 output = get_actions_with_access_level(db_session, serv, level)
                 results.extend(output)
-            print(yaml.dump(results)) if fmt == "yaml" else [print(result) for result in results]
+            print(yaml.dump(results)) if fmt == "yaml" else [
+                print(result) for result in results]
         # Get a list of all services in the database
         else:
             print("All services in the database:\n")
-            print(yaml.dump(all_services)) if fmt == "yaml" else [print(item) for item in all_services]
+            print(yaml.dump(all_services)) if fmt == "yaml" else [
+                print(item) for item in all_services]
     elif name is None and access_level and not wildcard_only:
-        print(f"All IAM actions under the {service} service that have the access level {access_level}:")
+        print(
+            f"All IAM actions under the {service} service that have the access level {access_level}:")
         level = transform_access_level_text(access_level)
         output = get_actions_with_access_level(db_session, service, level)
-        print(yaml.dump(output)) if fmt == "yaml" else [print(json.dumps(output, indent=4))]
+        print(yaml.dump(output)) if fmt == "yaml" else [
+            print(json.dumps(output, indent=4))]
     elif name is None and access_level and wildcard_only:
-        print(f"{service} {access_level.upper()} actions that must use wildcards in the resources block:")
-        output = get_actions_at_access_level_that_support_wildcard_arns_only(db_session, service, access_level)
-        print(yaml.dump(output)) if fmt == "yaml" else [print(json.dumps(output, indent=4))]
+        print(
+            f"{service} {access_level.upper()} actions that must use wildcards in the resources block:")
+        output = get_actions_at_access_level_that_support_wildcard_arns_only(
+            db_session, service, access_level)
+        print(yaml.dump(output)) if fmt == "yaml" else [
+            print(json.dumps(output, indent=4))]
     # Get a list of all IAM actions under the service that support the specified condition key.
     elif condition:
-        print(f"IAM actions under {service} service that support the {condition} condition only:")
-        output = get_actions_matching_condition_key(db_session, service, condition)
-        print(yaml.dump(output)) if fmt == "yaml" else [print(json.dumps(output, indent=4))]
+        print(
+            f"IAM actions under {service} service that support the {condition} condition only:")
+        output = get_actions_matching_condition_key(
+            db_session, service, condition)
+        print(yaml.dump(output)) if fmt == "yaml" else [
+            print(json.dumps(output, indent=4))]
     # Get a list of IAM Actions under the service that only support resources = "*"
     # (i.e., you cannot restrict it according to ARN)
     elif wildcard_only:
-        print(f"IAM actions under {service} service that support wildcard resource values only:")
+        print(
+            f"IAM actions under {service} service that support wildcard resource values only:")
         output = get_actions_that_support_wildcard_arns_only(
             db_session, service)
-        print(yaml.dump(output)) if fmt == "yaml" else [print(json.dumps(output, indent=4))]
+        print(yaml.dump(output)) if fmt == "yaml" else [
+            print(json.dumps(output, indent=4))]
     elif name and access_level is None:
         output = get_action_data(db_session, service, name)
-        print(yaml.dump(output)) if fmt == "yaml" else [print(json.dumps(output, indent=4))]
+        print(yaml.dump(output)) if fmt == "yaml" else [
+            print(json.dumps(output, indent=4))]
     else:
         # Get a list of all IAM Actions available to the service
         action_list = get_actions_for_service(db_session, service)
         print(f"ALL {service} actions:")
-        print(yaml.dump(action_list)) if fmt == "yaml" else [print(item) for item in action_list]
+        print(yaml.dump(action_list)) if fmt == "yaml" else [
+            print(item) for item in action_list]
 
 
 @query.command(
@@ -146,16 +160,19 @@ def arn_table(name, service, list_arn_types, fmt):
     # Get a list of all RAW ARN formats available through the service.
     if name is None and list_arn_types is False:
         raw_arns = get_raw_arns_for_service(db_session, service)
-        print(yaml.dump(raw_arns)) if fmt == "yaml" else [print(item) for item in raw_arns]
+        print(yaml.dump(raw_arns)) if fmt == "yaml" else [
+            print(item) for item in raw_arns]
     # Get a list of all the ARN types per service, paired with the RAW ARNs
     elif name is None and list_arn_types:
         output = get_arn_types_for_service(db_session, service)
-        print(yaml.dump(output)) if fmt == "yaml" else [print(json.dumps(output, indent=4))]
+        print(yaml.dump(output)) if fmt == "yaml" else [
+            print(json.dumps(output, indent=4))]
     # Get the raw ARN format for the `cloud9` service with the short name
     # `environment`
     else:
         output = get_arn_type_details(db_session, service, name)
-        print(yaml.dump(output)) if fmt == "yaml" else [print(json.dumps(output, indent=4))]
+        print(yaml.dump(output)) if fmt == "yaml" else [
+            print(json.dumps(output, indent=4))]
 
 
 @query.command(
@@ -186,8 +203,10 @@ def condition_table(name, service, fmt):
     # Get a list of all condition keys available to the service
     if name is None:
         results = get_condition_keys_for_service(db_session, service)
-        print(yaml.dump(results)) if fmt == "yaml" else [print(item) for item in results]
+        print(yaml.dump(results)) if fmt == "yaml" else [
+            print(item) for item in results]
     # Get details on the specific condition key
     else:
         output = get_condition_key_details(db_session, service, name)
-        print(yaml.dump(output)) if fmt == "yaml" else [print(json.dumps(output, indent=4))]
+        print(yaml.dump(output)) if fmt == "yaml" else [
+            print(json.dumps(output, indent=4))]