+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 2020-01-15
+### Added
+* `get_actions_matching_condition_key_crud_and_arn` is available. This provides some scaffolding for #21
+
 ## 2020-01-10
 ### Added
 * `get_all_actions_with_access_level` now supports "all" as a valid input, so you can easily request all IAM actions that are at a certain access level, regardless of service.
@@ -12,7 +16,6 @@
   - `tag` is now `tagging` to avoid inconsistency when writing templates
 * `command/write_policy/write_policy_with_actions` and `command/write_policy/write_policy_with_access_levels` can be called directly.
 * `writing/template/` now exposes `get_crud_template_dict` and `get_actions_template_dict` so developers can create the templates by calling the library. We might add on additional ones so they can just pass in lists without having to know the format, but not right now. They can pass that into `write_policy_with_actions` and `write_policy_with_access_levels`
-
 
 ## 2020-01-08
 * Fixed an issue where the bundled database was being appended to instead of overwritten when going through an update.